+++ web-api/src/main/java/statikk/webapi/controller/TeamBuilderController.java
@@ -5,7 +5,6 @@
  */
 package statikk.webapi.controller;
 
-import java.util.List;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,9 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-import statikk.domain.riotapi.model.ChampionMasteryDto;
-import statikk.domain.riotapi.model.Region;
-import statikk.domain.riotapi.model.SummonerDto;
 import statikk.webapi.model.ChampionSuggestion;
 import statikk.webapi.model.TeamBuilderProgressData;
 import statikk.webapi.service.TeamBuilderService;
@@ -33,7 +29,7 @@ public class TeamBuilderController extends BaseController {
 
     @ResponseBody
     @RequestMapping(value = "/suggestions", method = RequestMethod.POST)
-    public Map<Long, ChampionSuggestion> getSuggestions(@RequestBody TeamBuilderProgressData teamBuilderProgess) {
+    public Map<Integer, ChampionSuggestion> getSuggestions(@RequestBody TeamBuilderProgressData teamBuilderProgess) {
         return teamBuilderService.getChampionSuggestions(teamBuilderProgess);
     }
 }
